Add files via upload
A simple model for a market and the value of some commodity which varies randomly based on a set of weighted events. An "ifvestor", if statement for buying and selling a quantity, is currently employed by averaging the last couple prices and making a choice to sell or buy based on that.